Initial Firefox add on implementation and build environment.
Sets up an initial Firefox add on that can be run with jpm after it is built.
The add on has two built in commands that it will respond to:
commands - Returns a comma delimited set of commands that the bot will recognize.
bot_time - Returns the current time of the JS environment running the add on
resolves #1

Also includes the start of a build environment using Gradle. After installing
Gradle, run gradle tasks in the repository root directory to get available
tasks.
resolves #2