Locally unrecognized commands should be sent to a back end
If a command comes in that the local relay does not recognize then it
sends the command and it arguments to an external bot. The relay assumes
that the bot is running on localhost:8089 with a set REST API. To account for
the async nature of calling to an external program command processing has been
refactored to use JQuery deferreds.
Resolves #10

Reviewed by @mcw8d. Closes #18